use updated load function
